requirements.txt
reclaimer
pycollada
numpy
scipy
